Refactoring of scheduler and fiber work (#220)
* Quick node-based version of FS for console-based scenarios

* Removal of obviously unused files

* Renamed scamper-instance to scamper

---------
