+++ src/docs/index.ts
@@ -1,6 +1,9 @@
 import { createApp } from "vue"
-import { throwNull } from "../util.js"
 import DocsApp from "./DocsApp.vue"
+
+function throwNull(message: string): never {
+  throw new Error(message)
+}
 
 createApp(DocsApp).mount(
   document.getElementById("app") ?? throwNull("no app element"),
