--- src/index.ts
@@ -1,4 +1,0 @@
-// export { Scamper } from "./scamper.js"
-// export { ScamperVue } from "./scamper-vue.js"
-export { builtinLibs } from "./lib/index.js"
-// export { Prelude } from './lib/prelude/index.js'
