--- src/scamper-vue.ts
@@ -1,68 +1,0 @@
-// import * as Scheme from "./scheme"
-// import * as LPM from "./lpm"
-// import builtinLibs from "./lib"
-// import { VueDisplay } from "./lpm/output/vue"
-//
-// export class ScamperVue {
-//   display: VueDisplay
-//   prog: LPM.Prog | undefined
-//   machine: LPM.Thread | undefined
-//
-//   constructor(display: VueDisplay, src: string, isTracing = false) {
-//     this.display = display
-//     this.prog = Scheme.compile(this.display, src)
-//     if (this.prog) {
-//       const opts = LPM.cloneOptions(LPM.defaultOptions)
-//       opts.isTracing = isTracing
-//       this.machine = new LPM.Thread(
-//         "##main##",
-//         Scheme.mkInitialEnv(),
-//         this.prog,
-//         opts,
-//         builtinLibs,
-//         this.display,
-//         this.display,
-//         new Map([["scheme", Scheme.raiser]]),
-//       )
-//     }
-//   }
-//
-//   send(v: LPM.Value) {
-//     this.display.send(v)
-//   }
-//
-//   report(err: LPM.ScamperError) {
-//     this.display.report(err)
-//   }
-//
-//   async runProgram() {
-//     if (this.machine) {
-//       this.machine.cancelled = false
-//       await this.machine.evaluateAsync()
-//     }
-//   }
-//
-//   cancel() {
-//     // console.debug("attempted to cancel")
-//     if (this.machine) {
-//       this.machine.cancel()
-//     }
-//   }
-//
-//   runnerTree() {
-//     // TODO: need to update!
-//     // this.parseroutput.ast.renderTree(this.display, this.parseroutput.ast.nodes);
-//   }
-//
-//   stepProgram() {
-//     if (this.machine) {
-//       this.machine.step()
-//     }
-//   }
-//
-//   async stepStmtProgram() {
-//     if (this.machine) {
-//       await this.machine.stepExpr()
-//     }
-//   }
-// }
