--- src/util.ts
@@ -1,3 +1,0 @@
-export function throwNull(message: string): never {
-  throw new Error(message)
-}
