+++ src/web/components/use-scamper-session.ts
@@ -1,5 +1,5 @@
 import { computed, inject, type InjectionKey, provide, reactive, ref, type ShallowRef } from "vue"
-import { type DisplayRequest, ScamperInstance } from "../../scamper-instance"
+import { type DisplayRequest, ScamperInstance } from "../../scamper"
 import { SimpleErrorChannel } from "../../lpm/output/simple-error"
 import type { SchedulerId } from "../../scheduler"
 import type { ResultsPaneType } from "./use-results-pane"
